Include up to Jeff's commit: https://github.com/microsoft/onnxruntime/commit/9a4d1c772094ba2518b39d77207a2f852a9fedcb

git pull upstream 9a4d1c772094ba2518b39d77207a2f852a9fedcb
// where upstream == https://github.com/microsoft/onnxruntime.git

With this, `.\build.bat --build --skip_tests --config RelWithDebInfo --build_shared_lib --parallel --use_dml` now works again (because it finds the latest DirectML.h with the new enums). However, the ONNX conformance tests fail with E_NOTIMPL now inside `WINRT_RoGetActivationFactory` / `winrt::Windows::Foundation::IActivationFactory::QueryInterface()`.

+++ onnxruntime/test/providers/cpu/tensor/onehot_op_test.cc
@@ -283,7 +283,7 @@ TEST(OneHotOpTest, DefaultAxis_int32_MLFloat16_int32 /*indices, output, depth*/)
 
   std::vector<float> values{0.0f, 1.0f};
   std::vector<MLFloat16> fp16_values(values.size());
-  ConvertFloatToMLFloat16(values.data(), fp16_values.data(), values.size());
+  ConvertFloatToMLFloat16(values.data(), fp16_values.data(), static_cast<int>(values.size()));
 
   std::vector<float> output{0.0f, 1.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f,
                             0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 1.0f,
@@ -292,7 +292,7 @@ TEST(OneHotOpTest, DefaultAxis_int32_MLFloat16_int32 /*indices, output, depth*/)
                             0.0f, 0.0f, 0.0f, 0.0f, 1.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f,
                             0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 1.0f, 0.0f, 0.0f, 0.0f};
   std::vector<MLFloat16> fp16_output(output.size());
-  ConvertFloatToMLFloat16(output.data(), fp16_output.data(), output.size());
+  ConvertFloatToMLFloat16(output.data(), fp16_output.data(), static_cast<int>(output.size()));
 
   test.AddInput<int32_t>("indices", {2, 3}, {1, 9, 8, 2, 4, 6});
   test.AddInput<int32_t>("depth", {1}, {10});
